Fix regression that removed `addon` global.
- Write tests to catch such regressions in a future.
