+++ test/addons/privileged-panel/data/index.html
@@ -1,0 +1,3 @@
+<script>
+  addon.postMessage("hello addon")
+</script>
