+++ test/addons/privileged-panel/package.json
@@ -1,0 +1,3 @@
+{
+  "id": "test-privileged-addon"
+}
